double

+++ src/main/Row.java
@@ -3,7 +3,7 @@ package main;
 import java.util.HashMap;
 
 public class Row {
-    public HashMap<Integer,Integer> mapRow;
+    public HashMap<Integer,Double> mapRow;
     public int index;
 
     public Row(){
